Refactored GPhL UI to us JSON schemas - Improved UI. Still to be tested
